requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
colorama==0.4.6
discord==2.3.2
discord.py==2.5.2
filelock==3.13.1
frozenlist==1.6.0
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.3
networkx==3.3
numpy==2.1.2
pillow==11.0.0
propcache==0.3.1
sympy==1.13.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
typing_extensions==4.9.0
yarl==1.20.0
